style(scoring): match constant naming and keep Go test filenames
Rename the stack-buffer size to TOKEN_STACK_BUF_SIZE and move the
boundary tests into evaluator_test.go so Go keeps the required _test.go
suffix instead of a kebab-case test file.
